Moved file search and read functionality into sperate file_reader object
