+++ lib/src/draw_config_context.dart
@@ -3,11 +3,9 @@
 /// Use of this source code is governed by a BSD-style license that
 /// can be found in the LICENSE file.
 
-import 'dart:io';
-
 import 'package:ini/ini.dart';
-import 'package:path/path.dart' as path;
 
+import 'config_file_reader.dart';
 import 'exceptions.dart';
 
 const String kAccessToken = 'access_token';
@@ -24,13 +22,9 @@ const String kDefaultRedditUrl = 'https://www.reddit.com';
 const String kDefaultRevokeToken =
     r'https://www.reddit.com/api/v1/revoke_token';
 const String kDefaultShortUrl = 'https://redd.it';
-const String kFileName = 'draw.ini';
 const String kHttpProxy = 'http_proxy';
 const String kHttpsProxy = 'https_proxy';
 const String kKind = 'kind';
-const String kLinuxEnvVar = 'XDG_CONFIG_HOME';
-const String kLinuxHomeEnvVar = 'HOME';
-const String kMacEnvVar = 'HOME';
 const String kMessage = 'message_kind';
 const String kOauthUrl = 'oauth_url';
 const String kOptionalField = 'optional_field';
@@ -47,7 +41,6 @@ const String kSubmission = 'submission_kind';
 const String kSubreddit = 'subreddit_kind';
 const String kUserAgent = 'user_agent';
 const String kUsername = 'username';
-const String kWindowsEnvVar = 'APPDATA';
 
 final kNotSet = null;
 
@@ -91,10 +84,6 @@ class DRAWConfigContext {
     ],
     kRequiredField: const [kOauthUrl, kRedditUrl]
   };
-
-  /// Path to Local, User, Global Configuration Files, with matching precedence.
-  Uri _localConfigPath;
-  Uri _userConfigPath;
 
   Config _customConfig;
 
@@ -184,11 +173,10 @@ class DRAWConfigContext {
     _authorizeUrl = authorizeUrl ?? kNotSet;
     _userAgent = userAgent ?? kNotSet;
 
-    // Initialize Paths.
-    _localConfigPath = _getLocalConfigPath();
-    _userConfigPath = _getUserConfigPath();
+    final configFileReader = new ConfigFileReader(_configUrl);
+
     // Load the first file found in order of path preference.
-    final primaryFile = _loadCorrectFile();
+    final primaryFile = configFileReader.loadCorrectFile();
     try {
       _customConfig = new Config.fromStrings(primaryFile.readAsLinesSync());
     } catch (exception) {
@@ -196,29 +184,6 @@ class DRAWConfigContext {
     }
     // Load values found in the ini file, into the object fields.
     fieldMap.forEach((key, value) => _fieldInitializer(key, value));
-  }
-
-  /// Loads file from [_localConfigPath] or [_userConfigPath].
-  File _loadCorrectFile() {
-    if (_configUrl != null) {
-      final primaryFile = new File(_configUrl);
-      if (primaryFile.existsSync()) {
-        return primaryFile;
-      }
-    }
-    // Check if file exists locally.
-    var primaryFile = new File(_localConfigPath.toString());
-    if (primaryFile.existsSync()) {
-      _configUrl = _localConfigPath.toString();
-      return primaryFile;
-    }
-    // Check if file exists in user directory.
-    primaryFile = new File(_userConfigPath.toString());
-    if (primaryFile.existsSync()) {
-      _configUrl = _userConfigPath.toString();
-      return primaryFile;
-    }
-    throw new DRAWClientError('$kFileName, does not exist');
   }
 
   /// Take in the [type] which reflects the key in the [fieldMap]
@@ -339,36 +304,4 @@ class DRAWConfigContext {
   /// [key] is the key to be searched in the draw.ini file.
   String _fetchOrNotSet(final key) =>
       (_fetchOptional(key) ?? _fetchDefault(key) ?? kNotSet);
-
-  /// Returns path to user level configuration file.
-  /// Special Behaviour: if User Config Enviroment var unset, uses [$HOME] or the corresponding root path for the os.
-  Uri _getUserConfigPath() {
-    final environment = Platform.environment;
-    String osConfigPath;
-    // Load correct path for user level configuration paths based on operating system.
-    if (Platform.isMacOS) {
-      osConfigPath = path.join(environment[kMacEnvVar], '.config');
-    } else if (Platform.isLinux) {
-      osConfigPath = environment[kLinuxEnvVar] ??
-          path.join(environment[kLinuxHomeEnvVar], '.config');
-    } else if (Platform.isWindows) {
-      osConfigPath = environment[kWindowsEnvVar];
-    } else {
-      throw new DRAWInternalError('OS not Recognized by DRAW');
-    }
-    if (osConfigPath == null) {
-      // Sets osConfigPath to the corresponding root path based on the os.
-      final path.Context osDir = new path.Context();
-      final cwd = osDir.current;
-      osConfigPath = osDir.rootPrefix(cwd);
-    }
-    return Uri.parse(path.join(osConfigPath, kFileName));
-  }
-
-  /// Returns path to local configuration file.
-  Uri _getLocalConfigPath() {
-    final path.Context osDir = new path.Context();
-    final cwd = osDir.current;
-    return Uri.parse(path.join(cwd, kFileName));
-  }
 }
